refactor(snapshot): make the stele restore its own bootstrap command
`bootstrap snapshot` is left exactly as it was — `git diff origin/main` on
`snapshot.rs` is empty — and the stele restore moves out into a sibling,
`bootstrap stelae --source file://DIR`.

The two share a goal and nothing else: one unpacks a gzip tar of the storage
engines' own files over the data directory, the other reads deterministic CBOR
layers through the store traits. Their flags, failure modes and trust stories
all differ, so carrying both under one command produced options that only made
sense in pairs. Side by side they each read on their own, and the tarball path
stays untouched until Phase 6 decides its future.

The interactive bootstrap chooser lists both — a subcommand missing from it
would leave the menu silently incomplete — and prompts for the stele directory.
`stelae::run` takes no `Feedback`, which is the already-reported absence of
progress reporting rather than a new gap.

Nothing about the restore driver changed; this is CLI surface only.
